move ruler into views
